requirements.txt
os
pandas
numpy
pyyaml
sklearn
scikit-image
scikit-learn
statsmodels
webcolors
matplotlib
Pillow
nibabel